+++ dialogs/filing.py
@@ -2,6 +2,8 @@
 
 __author__ = 'Jero'
 
+import logging
+
 from PyQt5.QtWidgets import (QLabel, QLineEdit, QDialog, QVBoxLayout, QHBoxLayout, QGridLayout, QComboBox,
                              QTextEdit, QPushButton, QGroupBox)
 from PyQt5.QtGui import QTextDocument
@@ -128,7 +130,7 @@ class Filing(QDialog):  # Main dialog for filing template
     def on_clientComboBox_activated(self):
         """" Event handler for the client dropdown list """
 
-        print("[BET]: You selected", self.clientComboBox.currentText())
+        logging.info("[BET]: You selected {}".format(self.clientComboBox.currentText()))
         if self.clientComboBox.currentText() == 'GE':
             self.DEFAULT_SI = GE_DEFAULT
         elif self.clientComboBox.currentText() == 'Unilever':
@@ -162,7 +164,7 @@ class Filing(QDialog):  # Main dialog for filing template
         # Enable self.addButton
         self.addButton.setEnabled(True)
 
-        print("[BET]: Preview button clicked under Filing form")
+        logging.info("[BET]: Preview button clicked under Filing form")
 
     # OVERRIDING: starts here
     def accept(self):